Added extensions on IServiceCollection
